pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Project fot search differencies between two JSON files0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.9.2",
    "pytest-cov>=6.2.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[dependency-groups]
dev = [
    "ruff>=0.11.8",
    "pytest>=8.0.0",
    "coverage>=7.9.2",
]
